requirements.txt
PySide6>=6.6.0
google-generativeai>=0.8.3
anthropic>=0.34.0
toon-format>=0.9.0b1
Markdown>=3.5.0
pymdown-extensions>=10.7.0
pygments>=2.17.0
mss>=9.0.0
sounddevice>=0.4.6
soundfile>=0.12.1
pyaudiowpatch>=0.2.12.4
cryptography>=42.0.0
pypdf>=4.0.0
python-docx>=1.1.0
pydub>=0.25.1
Pillow>=10.2.0
python-dotenv>=1.0.0
